# [DDO-2166] Sign and notarize macOS releases + extras!