--- a/Dockerfile
+++ b/Dockerfile
@@ -5,12 +5,19 @@ ARG ALPINE_IMAGE_VERSION='3.14'
 #
 FROM alpine:${ALPINE_IMAGE_VERSION}
 
+ARG THELMA_LINUX_RELEASE
+
+COPY output/releases/${THELMA_LINUX_RELEASE} .
+
 # OS updates for security
 RUN apk update
 RUN apk upgrade
 
-# Copy Thelma into runtime image
-COPY output/release-assembly /thelma
+# Unpack Thelma into runtime image
+RUN mkdir /thelma && tar -xvf ${THELMA_LINUX_RELEASE} -C /thelma
+
+# Remove the copied tarball
+RUN rm ${THELMA_LINUX_RELEASE}
 
 ENV PATH="/thelma/bin:${PATH}"