--- a/scripts/sign-and-notarize.sh
+++ b/scripts/sign-and-notarize.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/zsh
 
 set -eo pipefail