--- a/scripts/sign-and-notarize.sh
+++ b/scripts/sign-and-notarize.sh
@@ -49,28 +49,23 @@ CMD_AUTH_FLAGS="--apple-id ${APPLE_ID} --password ${THELMA_MACOS_APP_PWD} --team
 create_keychain() {
 	# Decode the signing cert
 	local _cert_file="${WORKING_DIR}"/certificate.p12
-	echo "Writing cert ${_cert_file}"
 	echo "${THELMA_MACOS_CERT}" | base64 -d > "${_cert_file}"
 
 	# Create a temp keychain in the working dir
 	local KEYCHAIN_FILE="${WORKING_DIR}"/release.keychain
 	local _temp_keychain_pwd=temp-kc-pwd
-	echo "Creating keychain ${KEYCHAIN_FILE}"
 	security create-keychain -p ${_temp_keychain_pwd} "${KEYCHAIN_FILE}" # 2>&1 > /dev/null
 
 	# Setting default keychain
 	security default-keychain -s "${KEYCHAIN_FILE}" 2>&1 > /dev/null
 
 	# Unlock the keychain
-	echo "Unlocking keychain ${KEYCHAIN_FILE}"
 	security unlock-keychain -p ${_temp_keychain_pwd} "${KEYCHAIN_FILE}" 2>&1 > /dev/null
 
 	# Add the cert to the keychain
-	echo "Importing cert"
 	security import "${_cert_file}" -k "${KEYCHAIN_FILE}" -P "${THELMA_MACOS_CERT_PWD}" -T /usr/bin/codesign 2>&1 > /dev/null
 
 	# Allow codesign to use the keychain without a password prompt
-	echo "Setting partition list"
 	security set-key-partition-list -S apple-tool:,apple:,codesign: -s -k ${_temp_keychain_pwd} "${KEYCHAIN_FILE}" 2>&1 > /dev/null
 
 	echo "${KEYCHAIN_FILE}"
@@ -213,6 +208,7 @@ mkdir -p "${untar_dir}"
 tar -xf "${RELEASE_TARBALL}" -C "${untar_dir}"
 
 # Create the temp keychain
+echo "Creating keychain ${KEYCHAIN_FILE}"
 create_keychain
 
 # Sign each binary
@@ -248,6 +244,7 @@ tar -C "${untar_dir}" -czf "${RELEASE_TARBALL}" .
 echo "done"
 
 # Delete temp keychain
+echo "Removing keychain ${KEYCHAIN_FILE}"
 delete_keychain
 
 # Remove working dir