This PR adds signing and notarizing for macOS builds. It also adds the following goodies:
* Go deps are now cached between builds
* Builds are parellelized
* The docker build uses the linux release artifact instead of doing another build

Overall this speeds up builds a lot and minimizes the time impact that adding signing/notarizing would otherwise have.

Some notes on time improvements/speed ups, using [this](https://github.com/broadinstitute/thelma/runs/7362013023?check_suite_focus=true) before example and [this](https://github.com/broadinstitute/thelma/actions/runs/2707986760) after example:
* Build times
  * Before: 5m 53s
  * **After: 1m59s**
* Signing and notarizing:
  * Before: N/A
  * After: 4m 35s
* Image gen + push + release push:
  * Before: 3m 43s
  * **After: 1m 47s**
* Total build GHA run times:
  * Before: 10m 0s
  * After: 10m 0s

TL;DR: Signing and notarizing for macOS releases now works and it adds basically no time to the build workflow despite adding a step that takes 4.5mins.